fix deepsource code lint
