+++ pkg/mock/aws/CloudWatchLogsAPI.go
