+++ provider/aws/app_test.go
@@ -8,9 +8,9 @@ import (
 	awssdk "github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ecr"
 	"github.com/convox/convox/pkg/atom"
+	"github.com/convox/convox/pkg/mock/aws"
 	"github.com/convox/convox/pkg/structs"
 	"github.com/convox/convox/provider/aws"
-	"github.com/convox/convox/mock/aws"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
