+++ provider/aws/object_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/convox/convox/provider/aws"
-	"github.com/convox/convox/mock/aws"
+	"github.com/convox/convox/pkg/mock/aws"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
